+++ src/dune_rules/workspace.ml
@@ -56,6 +56,7 @@ module Context = struct
       ; fdo_target_exe : Path.t option
       ; dynamically_linked_foreign_archives : bool
       ; instrument_with : Lib_name.t list
+      ; merlin : bool
       }
 
     let to_dyn = Dyn.Encoder.opaque
@@ -72,6 +73,7 @@ module Context = struct
         ; fdo_target_exe
         ; dynamically_linked_foreign_archives
         ; instrument_with
+        ; merlin
         } t =
       Profile.equal profile t.profile
       && List.equal Target.equal targets t.targets
@@ -86,6 +88,7 @@ module Context = struct
       && Bool.equal dynamically_linked_foreign_archives
            t.dynamically_linked_foreign_archives
       && List.equal Lib_name.equal instrument_with t.instrument_with
+      && Bool.equal merlin t.merlin
 
     let fdo_suffix t =
       match t.fdo_target_exe with
@@ -145,7 +148,8 @@ module Context = struct
       and+ instrument_with =
         field_o "instrument_with"
           (Dune_lang.Syntax.since syntax (2, 7) >>> repeat Lib_name.decode)
-      and+ loc = loc in
+      and+ loc = loc
+      and+ merlin = field_b "merlin" in
       fun ~profile_default ~instrument_with_default ->
         let profile = Option.value profile ~default:profile_default in
         let instrument_with =
@@ -171,6 +175,7 @@ module Context = struct
         ; fdo_target_exe
         ; dynamically_linked_foreign_archives
         ; instrument_with
+        ; merlin
         }
   end
 
@@ -179,29 +184,25 @@ module Context = struct
       { base : Common.t
       ; switch : string
       ; root : string option
-      ; merlin : bool
       }
 
-    let to_dyn { base; switch; root; merlin } =
+    let to_dyn { base; switch; root } =
       let open Dyn.Encoder in
       record
         [ ("base", Common.to_dyn base)
         ; ("switch", string switch)
         ; ("root", option string root)
-        ; ("merlin", bool merlin)
         ]
 
-    let equal { base; switch; root; merlin } t =
+    let equal { base; switch; root } t =
       Common.equal base t.base
       && String.equal switch t.switch
       && Option.equal String.equal root t.root
-      && Bool.equal merlin t.merlin
 
     let t =
       let+ loc_switch, switch = field "switch" (located string)
       and+ name = field_o "name" Context_name.decode
       and+ root = field_o "root" string
-      and+ merlin = field_b "merlin"
       and+ base = Common.t in
       fun ~profile_default ~instrument_with_default ~x ->
         let base = base ~profile_default ~instrument_with_default in
@@ -221,7 +222,7 @@ module Context = struct
                 ])
         in
         let base = { base with targets = Target.add base.targets x; name } in
-        { base; switch; root; merlin }
+        { base; switch; root }
   end
 
   module Default = struct
@@ -323,6 +324,7 @@ module Context = struct
       ; fdo_target_exe = None
       ; dynamically_linked_foreign_archives = true
       ; instrument_with = Option.value instrument_with ~default:[]
+      ; merlin = false
       }
 
   let build_contexts t =
@@ -539,10 +541,13 @@ let step1 clflags =
                Context_name.Set.union !defined_names
                  (Context_name.Set.of_list (Context.all_names ctx));
              match (ctx, acc) with
-             | Opam { merlin = true; _ }, Some _ ->
+             | Opam { base = { merlin = true; _ }; _ }, Some _
+             | Default { merlin = true; _ }, Some _ ->
                User_error.raise ~loc:(Context.loc ctx)
                  [ Pp.text "you can only have one context for merlin" ]
-             | Opam { merlin = true; _ }, None -> Some name
+             | Opam { base = { merlin = true; _ }; _ }, None
+             | Default { merlin = true; _ }, None ->
+               Some name
              | _ -> acc)
        in
        let contexts =